Refactor extract testable function to a header file
The functions to be tested must be includable independently of
the main function.
