+++ minunit/src/count_cards.c
@@ -7,8 +7,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-/* Returns the card value */
-int card_value(char card_name[]);
+#include "count_cards.h"
 
 int main() {
 
@@ -20,23 +19,4 @@ int main() {
     val = card_value(card_name);
     printf("The card value is %i\n", val);
     return 0;
-}
-
-int card_value(char card_name[]) {
-
-    int val = 0;
-
-    if (card_name[0] == 'K') {
-        val = 10;
-    } else if (card_name[0] == 'Q') {
-        val = 10;
-    } else if (card_name[0] == 'J') {
-        val = 10;
-    } else if (card_name[0] == 'A') {
-        val = 11;
-    } else {
-        val = atoi(card_name);
-    }
-
-    return val;
 }
